Refactor Queue into an abstract base class and a concrete Redis implementation
