+++ saq/queue/__init__.py
@@ -1,0 +1,8 @@
+from saq.queue.base import JobError, Queue
+from saq.queue.redis import RedisQueue
+
+__all__ = [
+    "JobError",
+    "Queue",
+    "RedisQueue",
+]
