+++ tests/helpers.py
@@ -1,12 +1,13 @@
 import typing as t
 
-from saq.queue import Queue
+from saq.queue import Queue, RedisQueue
 
 
-def create_queue(**kwargs: t.Any) -> Queue:
-    return Queue.from_url("redis://localhost:6379", **kwargs)
+def create_queue(**kwargs: t.Any) -> RedisQueue:
+    queue = t.cast(RedisQueue, Queue.from_url("redis://localhost:6379", **kwargs))
+    return queue
 
 
-async def cleanup_queue(queue: Queue) -> None:
+async def cleanup_queue(queue: RedisQueue) -> None:
     await queue.redis.flushdb()
     await queue.disconnect()
